// ironsworn-starforged/src/system/dataforged/types.ts
export declare interface GameDataRoot {
  $schema?: string | undefined
  'Asset Types': IAssetType[]
  Encounters: IEncounterStarforged[] | IEncounterNatureInfo[]
  'Move Categories': IMoveCategory[]
  'Oracle Categories': IOracleCategory[]
  'Setting Truths'?: ISettingTruth[]
}
// ...
export declare interface IAsset
  extends IHasId,
    IHasName,
    IHasDisplay,
    IHasSource,
    Partial<IHasAliases> {
  /**
   * @example "Starforged/Assets/Path/Bounty_Hunter"
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+/[A-z_-]+$
   */
  $id: string
  /**
   * The asset's name - the title printed on the card.
   * @example "Bounty Hunter"
   */
  Name: string
  Display: IDisplayWithTitle
  /**
   * Describes any states that the asset might have, such as "Broken". Some states may disable the asset entirely.
   */
  States?: IAssetState[] | undefined
  /**
   * The ID of the asset's parent AssetType
   * @example "Starforged/Assets/Path"
   */
  'Asset Type': IAssetType['$id']
  /**
   * Information on the asset's usage, such as whether its abilities are shared amongst the player characters.
   */
  Usage: IAssetUsage
  /**
   * Details on what attachments (other assets) are accepted by this asset.
   */
  Attachments?: IAssetAttachment | undefined
  /**
   * Data describing the Input controls that should be embedded in the card. Inputs embedded in specific asset abilities appear as keys of the corresponding ability object, instead.
   */
  Inputs?: (IInputText | IInputSelect)[] | undefined
  /**
   * An optional markdown string representing the requirement text that appears at the top of some asset cards.
   * @markdown
   * @example "If you wear your finely crafted set of personal armor..."
   */
  Requirement?: string | undefined
  /**
   * The asset's abilities.
   */
  Abilities: [IAssetAbility, IAssetAbility, IAssetAbility]
  /**
   * Information on this asset's condition meter, if any.
   */
  'Condition Meter'?: IConditionMeter | undefined
  Tags?: string[] | undefined
}

/**
 * Represents one of an asset's three abilities.
 * @public
 */
export declare interface IAssetAbility extends IHasId, IHasText {
  /**
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+/[A-z_-]+/Abilities/[1-3]$
   */
  $id: string
  /**
   * Ironsworn companion assets provide names for their abilities. Starforged asset abilities do not have names.
   */
  Name?: string | undefined
  /**
   * New moves added by this asset ability.
   */
  Moves?: IMove[] | undefined
  /**
   * User inputs (text, clocks, etc) associated with this asset ability.
   */
  Inputs?: (IInputNumber | IInputClock | IInputText)[] | undefined
  /**
   * Information on how this ability alters moves when enabled.
   */
  'Alter Moves'?: IAlterMove[] | undefined
  /**
   * Information on how this ability alters its parent asset when enabled.
   */
  'Alter Properties'?: IAssetAlterProperties | undefined
  /**
   * Information on how this ability alters its owner's momentum (triggers an effect on burn, on reset, etc)
   */
  'Alter Momentum'?: IAlterMomentum | undefined
  /**
   * Whether the asset ability is enabled or not. In most cases, the first asset ability defaults to 'true' and the others to 'false'. If none of an asset's abilities are set to 'true', the player can pick which the ability they start with when purchasing the asset.
   */
  Enabled: boolean
}

/**
 * Describes changes that an asset ability makes to its parent asset when active. Any properties with object values should be merged recursively.
 *
 * @example An `IAssetAlterProperties` that would set `IAsset["Condition Meter"].Max` to 3, and leave its other properties unchanged:
 * ```javascript
 * { "Condition Meter": { Max: 3 } }
 * ```
 * @public
 */
export declare interface IAssetAlterProperties extends PartialDeep<IAsset> {}

/**
 * Details which assets are valid attachments. The most prominent example in *Ironsworn: Starforged* is the STARSHIP asset (`Starship/Assets/Command_Vehicle/Starship`); Rover (`Starship/Assets/Support_Vehicle/Rover`) also has an elective ability that adds this property.
 * @public
 */
export declare interface IAssetAttachment {
  /**
   * The type of asset that this asset accepts as attachments.
   */
  'Asset Types': IAssetType['$id'][]
  /**
   * The maximum number of attached assets accepted by this asset. If undefined or null, there is no maximum.
   * @nullable
   */
  Max: number | null
}

/**
 * Describes a possible state for an asset, like the "Broken" status for certain assets (mainly Modules in *Starforged*).
 *
 * States are frequently toggled on and off by players; for real-world gameplay, this is generally represented by flipping the card over. A checkbox or other on/off toggle might serve the same function in a digital implementation.
 * @public
 */
export declare interface IAssetState extends IHasName {
  /**
   * A string label name or label for the state.
   * @example "Broken"
   */
  Name: string
  /**
   * Whether this state is currently enabled.
   */
  Enabled: boolean
  /**
   * Whether this state should disable the entire asset when `IAssetState.Enabled === true`
   */
  'Disables asset': boolean
  /**
   * Whether this state counts as an Impact for the asset's owner.
   *
   * Note that for vehicles, this shouldn't be applied automatically unless your implementation has some way of telling which vehicle the PC is currently using.
   */
  Impact: boolean
  /**
   * Whether or not this state is permanent.
   */
  Permanent: boolean
}

/**
 * Represents an Asset Type such as Command Vehicle, Companion, or Path, and serves as a container for all assets of that type.
 * @public
 */
export declare interface IAssetType
  extends IHasName,
    IHasId,
    IHasDescription,
    IHasDisplay,
    IHasSource,
    Partial<IHasAliases> {
  /**
   * @example "Ironsworn/Assets/Ritual"
   * @example "Starforged/Assets/Command_Vehicle"
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+$
   */
  $id: string
  /**
   * The assets that belong to this asset type.
   */
  Assets: IAsset[]
  /**
   * @example "Ritual"
   * @example "Command Vehicle"
   */
  Name: string
  Display: IDisplayWithTitle
  Usage: IAssetUsage
}

/**
 * @public
 */
export declare interface IAssetUsage {
  /**
   * Whether the asset's abilities are shared with Allies.
   *
   * If set to `true`, the asset's abilities can be invoked by **any** player character; if your app facilitates co-op or guided play, consider how you might expose these abilities to players other than the asset's owner.
   *
   * Defaults to `true` for Command Vehicle, Support Vehicle, and Module assets.
   */
  Shared: boolean
}
// ...
export declare interface IConditionMeter extends IMeterBase {
  /**
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+/[A-z_-]+/Condition_Meter$
   */
  $id: string
  Min: 0
  /**
   * The conditions that can apply to this meter.
   */
  Conditions: MeterCondition[]
  Aliases?: MeterAlias[] | undefined
}
// ...
export declare interface IDisplay {
  /**
   * A URL pointing to a single SVG icon.
   * @pattern ^\.\./\.\./img/vector/[A-z-_0-9/]+\.svg$
   */
  Icon?: string | undefined
  /**
   * An array of URLs pointing to one or more WEBP images.
   * @pattern ^\.\./\.\./img/raster/[A-z-_0-9/]+\.webp$
   */
  Images?: string[] | undefined
  /**
   * A hex color associated with this item, for use as e.g. an accent color in its display.
   * @pattern ^#[A-f0-9][A-f0-9][A-f0-9][A-f0-9][A-f0-9][A-f0-9]$
   */
  Color?: string | undefined
  /**
   * The title of this item as it appears printed in the rulebook. Intended for use as the item's header, label, etc.
   */
  Title?: string | undefined
}
// ...
export declare interface IDisplayWithTitle extends IDisplay {
  Title: string
}
// ...
export declare interface IEncounterStarforged extends IEncounter {
  /**
   * @pattern ^Starforged/Encounters/[A-z_-]+$
   */
  $id: string
  Nature: EncounterNatureStarforged
  Summary: string
  Variants: IEncounterVariant[]
}
// ...
export declare interface IHasAliases {
  /**
   * Alternate names for this item, including: names it had earlier in development that have since changed, alternate spellings/punctuation, common misspellings, and so on.
   */
  Aliases: string[]
}

/**
 * Interface for items with a user-facing markdown description, consisting of one or more paragraphs.
 * @public
 */
export declare interface IHasDescription {
  /**
   * A user-facing markdown description of the item, consisting of one or more paragraphs.
   * @markdown
   */
  Description: string
}

/**
 * Interface for items with rendering information.
 * @public
 */
export declare interface IHasDisplay {
  /**
   * Data relevant to this item's display/rendering.
   */
  Display: IDisplay
}
// ...
export declare interface IHasId {
  /**
   * The item's unique string ID.
   * @pattern ^(Starforged|Ironsworn)/[0-9A-z_/-]+$
   */
  $id: string
}

/**
 * Interface for items with a Name key.
 * @public
 */
export declare interface IHasName {
  /**
   * The item's internal name. Should be unique among its sibling elements, as this key is often used (along with the object's ancestors) to generate its $id.
   *
   * If the item has Display.Title, that should be preferred for most user-facing labels.
   */
  Name: string
}
// ...
export declare interface IHasSource {
  /**
   * Information on this item's source.
   */
  Source: ISource
}
// ...
export declare interface IHasText {
  /**
   * The item's rules text as a markdown string.
   * @markdown
   */
  Text: string
}

/**
 * A stub interface representing an input widget of any type.
 * @see {@link IInputNumber}, {@link IInputClock}, {@link IInputText}, {@link IInputSelect}
 * @public
 */
export declare interface IInput extends IHasId, IHasName {
  /**
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+/[A-z_-]+(/Abilities/[1-3])?/Inputs/[A-z_-]+$
   */
  $id: string
  'Input Type': InputType
  /**
   * Whether the input's value is expected to change over the course of a campaign. For example, name fields are typically `false`, while something like a clock or tally would be `true`.
   *
   * It's a good idea to make everything editable regardless, but this property might inform whether your UI presents that functionality "front and center" or as a secondary interaction (via long press, right click, etc);
   */
  Adjustable: boolean
}
// ...
export declare interface IInputNumber extends IInput {
  'Input Type': InputType.Number
  Min: number
  /**
   * @nullable
   */
  Max: number | null
  Step: 1
  Value: number
}
// ...
export declare interface IInputSelect extends IInput {
  'Input Type': InputType.Select
  /**
   * Hints which attribute(s) set by this dropdown's options.
   */
  Sets: IInputSelectAttributeDefinition[]
  Options: IInputSelectOption[]
}

/**
 * Provides hints for the keys and typing of an {@link IInputSelect}'s child {@link IInputSelectOption}s.
 * @typeParam V - The type(s) of the value(s) set by this item's options.
 * @public
 */
export declare interface IInputSelectAttributeDefinition {
  Key: string
  Type: InputSelectOptionType
}

/**
 * Represents an option in an {@link IInputSelect}.
 * @public
 */
export declare interface IInputSelectOption extends IHasId, IHasName {
  /**
   * @pattern ^(Starforged|Ironsworn)/Assets/[A-z_-]+/[A-z_-]+/Inputs/[A-z_-]+/Options/[A-z_-]+$
   */
  $id: string
  /**
   * A array describing what attribute keys should be set to when this option is active. *All* items in the array should be set in this manner.
   */
  Set: (
    | IInputSelectOptionSetterStat
    | IInputSelectOptionSetterMeter
    | IInputSelectOptionSetterNumber
    | IInputSelectOptionSetterString
  )[]
}
// ...
export declare interface IInputText extends IInput {
  'Input Type': InputType.Text
}
// ...
export declare interface ISettingTruth
  extends IHasId,
    IHasName,
    IHasSource,
    IHasDisplay,
    Partial<IHasSuggestions>,
    IHasTable {
  /**
   * @pattern ^(Starforged|Ironsworn)/Setting_Truths/[A-z_-]+$
   */
  $id: string
  Name: string
  /**
   * The 'canonical' options for this setting truth category.
   */
  Table: ISettingTruthOption[]
  /**
   * A Markdown version of the text that appears at the end of each Truth entry; it offers suggestions on the character's assets and background.
   * @markdown
   */
  Character: string
  Display: IDisplayWithTitle
}
// ...
export declare interface ISource {
  /**
   * The title of the source.
   *
   * For 'canonical' content, use one of the enumerated `SourceTitle` strings.
   *
   * For 3rd-party content (including homebrew) that's been released as part of a titled document, use the title of that document (e.g. "Steelforged", "Ironsmith").
   *
   * If the source has no particular title (for instance, it's a single custom element in a VTT implementation), use "Custom".
   */
  Title: SourceTitle | string
  /**
   * The author(s) of this item. For 'canonical' content, this one's usually pretty obvious 😉 However, it's included so that homebrew content can use the same interface/schema.
   * @default ["Shawn Tomkin"]
   */
  Authors: string[]
  /**
   * The 6-number date string formatted as `MMDDYY`. Relevant only during Starforged development; it will be deprecated once the game is released.
   * @pattern ^(0[1-9]|1[0-2])([0-2][1-9]|3[0-1])([0-9][0-9])$
   */
  Date?: string | undefined
  /**
   * The page on which the item appears most prominently in the source material (if it's in a format that uses page numbers).
   */
  Page?: number | undefined
  /**
   * The URL where the source material is available.
   * @pattern ^https?://.*$
   */
  Url?: string | undefined
}
// ...
export declare interface Starforged extends GameDataRoot {
  Encounters: IEncounterStarforged[]
  'Setting Truths': ISettingTruth[]
}

/**
 * @public
 */
export declare const starforged: Starforged